+++ test/Address.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace FBE\Test;
+
+/**
+ * Address data class (readonly, immutable)
+ */
+readonly class Address
+{
+    public function __construct(
+        public string $city,
+        public string $country
+    ) {}
+}
+